+++ project/Source/Component/MeshRenderer2D/MeshRenderer2D.h
@@ -114,7 +114,7 @@ public:
 	/// ‰æ‘œ‚Ì”¼“§–¾“x‚ðŽæ“¾
 	/// </summary>
 	/// <returns></returns>
-	float GetAlpha() const { return alpha; } 
+	int GetAlpha() const { return alpha; } 
 
 private:
 	int hImage;
